+++ projects/demo-application/src/app/demo.model.ts
@@ -1,6 +1,5 @@
 export class Demo {
   name: string = ''
-  birthday: Date = new Date();
   isActive: boolean = false;
   count: number = 0;
 }