+++ projects/query-search/src/lib/components/fields/name-field/name-field.component.ts
@@ -1,9 +1,9 @@
-import {Component, EventEmitter, Input, Output} from '@angular/core';
+import {ChangeDetectorRef, Component, EventEmitter, Input, Output} from '@angular/core';
 import {QuerySearchService} from "../../../query-search.service";
 import {BehaviorSubject, Observable} from "rxjs";
-import {QueryField} from "../../../models";
+import {QueryField, QueryItem} from "../../../models";
 import {map, startWith, switchMap} from "rxjs/operators";
-import {filterFieldNames, getFieldType} from "../../../helpers/query-search.helpers";
+import {filterFieldNames, getFieldType} from "../../../helpers/general.helpers";
 
 @Component({
   selector: 'name-field',
@@ -11,29 +11,33 @@ import {filterFieldNames, getFieldType} from "../../../helpers/query-search.help
   styleUrls: ['./name-field.component.scss']
 })
 export class NameFieldComponent {
-  @Output()
-  selectedFieldChange = new EventEmitter<QueryField>(true);
 
   @Input()
-  set selectedField(newValue: QueryField) {
+  set item(newValue: QueryItem) {
     if (!!newValue) {
-      this._selectedField = newValue;
-      this.fieldChanged(newValue, false);
+      this._item = newValue;
+      if (!!this._item.fieldName) {
+        this.fieldChanged(this._item.fieldName, false);
+      }
     }
   }
 
-  get selectedField() {
-    return this._selectedField;
+  get item() {
+    return this._item;
   }
+
+  @Output()
+  itemChange: EventEmitter<QueryItem> = new EventEmitter<QueryItem>();
 
   allFields: Observable<QueryField[]>;
   visibleFields: Observable<QueryField[]>;
   searchValue: BehaviorSubject<string> = new BehaviorSubject<string>('');
   nameFieldTrigger: string;
 
-  private _selectedField: QueryField;
+  private _item: QueryItem;
 
-  constructor(private querySearchService: QuerySearchService) {
+  constructor(private querySearchService: QuerySearchService,
+              public changeDetectorRef: ChangeDetectorRef) {
     this.allFields = querySearchService.fields;
     this.visibleFields = this.searchValue.pipe(
       startWith(null),
@@ -53,16 +57,31 @@ export class NameFieldComponent {
     this.searchValue.next(searchValue);
   }
 
-  fieldChanged(event: QueryField, emit = true) {
-    if (emit) {
-      this.selectedFieldChange.emit(event);
-    }
+  fieldChanged(event: string, emit = true) {
+    const field = this.querySearchService.getField(event);
 
-    this.querySearchService.log('Name field - fieldChanged', event);
-    if (this.querySearchService.showFieldNameSuffix) {
-      this.nameFieldTrigger = `<span class="field-name">${event.label || event.name}</span><span class="field-suffix">${getFieldType(event)}</span>`;
-    } else {
-      this.nameFieldTrigger = `<span class="field-name">${event.label || event.name}</span>`;
+    if (!!field) {
+      if (this.item.type !== field.type || this.item.fieldName !== field.name) {
+        this.item.value = null;
+      }
+
+      this.item.type = field.type;
+      this.item.fieldName = field.name;
+
+      if (emit) {
+        this.querySearchService.log('NameFieldComponent - Emitting Change:', field)
+        this.itemChange.emit(this.item);
+      } else {
+        this.querySearchService.log('NameFieldComponent - Not Emitting Change:', field);
+      }
+
+      if (this.querySearchService.showFieldNameSuffix) {
+        this.nameFieldTrigger = `<span class="field-name">${field.label || field.name}</span><span class="field-suffix">${getFieldType(field)}</span>`;
+      } else {
+        this.nameFieldTrigger = `<span class="field-name">${field.label || field.name}</span>`;
+      }
+
+      this.changeDetectorRef.detectChanges();
     }
   }
 }